WIP - moved plugs; set up a new token-protected route plug
